fix(nodes): move InvocationConfig to baseinvocation.py
